refactored - theme, removed duplicate files, refactored comp structure, refactored file structure, renamed files for better clarity
