+++ src/App.tsx
@@ -1,10 +1,11 @@
-import Header from './Header.tsx';
-import WorkPage from './projectComps/WorkPage.tsx';
-import HomeSplash from './HomeSplash.tsx';
-import ContactPage from './ContactPage.tsx';
+import Header from './components/layout/Header.tsx';
+import HomeSplash from './pages/Home.tsx';
+import ContactPage from './pages/Contact.tsx';
+import ProjectList from './components/projects/ProjectList.tsx';
+
 import { Routes, Route } from 'react-router-dom';
-import ThemeToggle from './ThemeButton.tsx';
-import { ThemeProvider } from './providor/ThemeContext.tsx';
+import ThemeToggle from './components/layout/ThemeButton.tsx';
+import { ThemeProvider } from './provider/ThemeContext.tsx';
 
 function App() {
   return (
@@ -14,7 +15,7 @@ function App() {
         <ThemeToggle>
           <Routes>
             <Route path="/" element={<HomeSplash />} />
-            <Route path="/work" element={<WorkPage />} />
+            <Route path="/work" element={<ProjectList />} />
             <Route path="/contact" element={<ContactPage />} />
           </Routes>
         </ThemeToggle>
