+++ src/pages/Contact.tsx
@@ -1,6 +1,6 @@
-import resume from '../src/assets/docs/JoelKram_Frontend Dev.pdf';
-import { Headshot } from '../src/threejs/Headshot.tsx';
-export default function ContactPage() {
+import resume from '../assets/docs/JoelKram_Junior Fullstack Developer.pdf';
+import { Headshot } from '../threejs/Headshot.tsx';
+export default function Contact() {
   return (
     <div className="min-h-screen bg-white dark:bg-zinc-700 transition-colors duration-300 ease-linear">
       <section className="max-w-4xl mx-auto px-4 py-8 text-black dark:text-white text-[11pt] font-['Neue'] tracking-[-0.3px]">
