--- src/projectComps/WorkPage.tsx
@@ -1,8 +1,0 @@
-import PageContainer from './PageContainer.tsx';
-export default function WorkPage() {
-  return (
-    <div>
-      <PageContainer />
-    </div>
-  );
-}
